Treat D-Bus properties named "Type" in special way.
Gdbus-codegen emits C function names with an underscore appended to
"Type" properties to avoid some name clash, so we need to do the same
when extracting Doxygen documentation.

Of course this little detail is not documented anywhere, of course we
ran straight into it, and of course the presence of two properties
named "Type" and "Type_" in the same interface would generate invalid
C code.

No kudos to GLib today.

Change-Id: I18d434ea0c3ea998cc6ba576265c291d8e9664fa