Header files for de.tahifi.Lists.Navigation error codes.
It is very helpful to be able to see the actual cause of any errors
returned by the lists-related D-Bus methods. This way the caller
roughly knows what's going on and can do something to help the user
resolving problems.

To make it work, appropriate error codes must be determined at the
point of failure, and these need to bubble all the way up to the D-Bus
adapter code (or whoever called). This can be done either by return
codes, output parameters, or exceptions, whatever fits best.

For C, a header file containing a plain enum and full documentation of
error codes is provided. For C++, a small class based on the C header
file is provided. A moderate amount of macro magic is invoked to make
the enum names acceptable for C and C++, respectively.

The Documentation in the XML file points to the header files and does
not attempt to document the error codes any further.

Change-Id: Iafc380b2cde84755c218edf237e59b305ed84725