Fix de.tahifi.Power.Client interface.
The IDs cannot be extracted from D-Bus callers because there is no
true point-to-point connection, at least not in a way that would be
usable for generating IDs from caller names. Instead, we now assign
numeric IDs to registered clients and expect clients to use that ID.
The string ID is only used at registration time.

Change-Id: I5bace08c652640dcac437c2a8cd9584ef6032058