Add stream ID to all de.tahifi.Streamplayer.Playback signals.
The stream ID is required to avoid certain race conditions. It is
possible for the NowPlaying signal being emitted before
MetaDataChanged signals for the previously playing stream have been
sent around. This is probably caused by the gapless playback code
implemented in playbin and is not a problem if the meta data is only
produced by the stream player.

If, however, meta data can also be pulled from sources other than the
stream itself, then clients must be able to tell which exact stream
the meta data posted around belongs to.

Also add documentation.

Change-Id: I89c9c85c338f9c0fcaacd612c7e2631a322a8955