Enable cover art management in cover art sources.
Cover art is identified by stream key, associating streams with their
respective cover art. The key is always provided by the sources of
cover art, i.e., parts of the system that extract the cover art. It is
their responsibility to generate stable stream keys that can
permanently identify streams (unlike stream IDs, which can only be
used to identify playing and queued streams).

Currently, sources of cover art are the various list navigation
implementations.

In addition, the stream player expects a stream key being passed when
queuing a stream. This is because, by definition, the stream player is
dumb and does not know a lot about the origin of streams. It cannot
generate stream keys, but it will use the passed stream key to push
out additional cover art extracted from streams while they are
playing. This makes the stream player an extra, delegated source of
cover art.

Change-Id: If3f67ac80a0f3dfe81b9fc640ba4450155ea8db7