Add list method for overriding caching mode (#616).
For playing a list of files directly from the directory hierarchy, it
is necessary to keep those lists in cache even if the default caching
policy does not allow this.

Note that list brokers may implement COW to keep cached copies around
for use by the player (which would continue using the ID forced into
cache for its playback root and never try to re-enter it), while
providing non-cached navigation to the list browser (which would at
some point try to re-enter the list, thus get a new ID). Instead of
purging subtrees on enter, the subtree forced into cache would be
detached for the main tree (sort of) and would still continue to work.
The API as it is designed now should already work for such an
implementation.

Change-Id: Icb3ec0e893153ff043b8578ade7e6821975c5af8