Add de.tahifi.Dcpd.Playback.Resume signal (#604).
Like de.tahifi.Dcpd.Playback.Play or de.tahifi.Dcpd.Playback.Stop, the
new signal for resuming playback from a previously stored location
relates to the active audio source, if any.

This is not a "smart" resume which would switch to the most recently
used audio source and try to resume from there. Users of this signal
are required to switch to any audio source they are interested in
before requesting resumption.

Change-Id: I8efae5610ee88f996a8a74ecbd628e06cc02e6e0