Pass arbitrary data with audio path requests (#895).
Requests for audio path switches and deactivation of current audio
path now require---or rather allow---passing extra data in form of a
dictionary. These data are passed to the functions defined by the
de.tahifi.AudioPath.Source and de.tahifi.AudioPath.Player interfaces.
This way, each part in the both, the released and the new audio path,
can see these data and react to them if they want to.

The dictionary of extra data may be left empty in case default
behaviors are good enough.

Change-Id: I78077ac9240bb1cefe55e21f256a4dacc3cbe180