Redefine de.tahifi.Lists.Navigation.RealizeLocation.
This commit is a precursor for more changes to come. RealizeLocation
has been changed to use a more general cookie mechanism which are
going to be used by more list navigation methods soon.

Instead of AbortRealizeLocation() for cancellation, we have
DataAbort() now. Instead of a RealizeLocationResult signal, we have
DataAvailable and DataError signals for client notification, and
RealizeLocationByCookie() for retrieval of results now.

Also, RealizeLocation() cannot return results directly anymore, but
only an error code and a cookie. The results must be fetched in a
succeeding call of RealizeLocationByCookie(). The client will be
informed if and when it should call this function by observing the new
DataAvailable signal (results can and should be fetched) and DataError
signal (no results available, no further calls necessary).

Change-Id: I3a23f3a7f4cbe9a5b7ba797e2d7fbe80ad4ee7d3