Streamplayer: Replace Paused signal by PauseState signal.
This is a breaking API change. The NowPlaying signal does NOT imply
end of pause state anymore as this information is now transmitted
using the PauseState signal.

The reason for this change is that sometimes GStreamer reports end of
pause state for newly started tracks, even in case the stream has
never been paused before. When the signal is delivered, it is
impossible for streamplayer to tell whether or not this is a valid
state change or, which is worse, which stream from our queue is
currently playing (because of gapless mode). Since we were using the
NowPlaying signal to report end of pause state, it is impossible for
streamplayer to emit a reliable NowPlaying signal in such a case,
leading to accidental skips in drcpd's queue.

Change-Id: I80645bb186f8aafb2b342a3c95883fa4232b3ca9