combine_iface_pages: Script for avoiding Doxygen warnings.
There is a pretty stupid "feature" in Doxygen which emits a warning
for multiple uses of \page with the same page identifier. It does the
right thing: merge the pages into a single page. And it emits a
warning, which is always annoying in CI.

According to the Doxggen issue tracker, this warning should have been
removed in recent versions, but it's still there. So we are doing it
manually just to work around the warning.

Change-Id: Ic8e17b6cee7049fdf03ac7a562910e07bb81578c