taddybus-mockgen: Allow signals to be checked by external function.
Unit tests may now implement any partial checks of D-Bus signals in
cases where the full check would be too detailed. It is also possible
to bypass the check altogether, turning such an expectation into a
"need any D-Bus signal of this kind here" type of check.

We can extend this to method calls as well, but we don't need such a
thing at the moment. To be added when needed.

Change-Id: Ic623b1954c8732d688b11428414c49cf98fca95c